Changed SparseDirectUMFPACK to use umfpack_dl_* routines
git-svn-id: https://svn.dealii.org/trunk@17243 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/include/lac/sparse_direct.h
@@ -1248,8 +1248,8 @@ class SparseDirectUMFPACK : public Subscriptor
                                       * The arrays in which we store the data
                                       * for the solver.
                                       */
-    std::vector<int> Ap;
-    std::vector<int> Ai;
+    std::vector<long int> Ap;
+    std::vector<long int> Ai;
     std::vector<double> Ax;
 
                                      /**
